introspection: Calculate introspection schema deterministically
Sort values of the a golang map such that we can calculate a
deterministic graphql introspection schema.

Previously, registerQuery() and registerType() methods were converting
values of a map into a slice by iterating in a for-range loop. Since
golang's map iterates in a random order, we ended up getting a different
Schema object every time.