reactive: add a defensive copy utility function
Using reactive.Cache can be tricky. Modifying values returned by
reactive.Cache causes all kind of problems, but it's hard to tell if
your code does or does not do since Go has no immutable types.

This commit introduces a utility function to take a defensive copy of a
value. We can't really make a copy of a value in go (because of reflect
limitations), but we can take a snapshot that we can later compare with
another snapshot.