graphql: add schemabuilder support for NonNull and update introspection test
Add schemabuilder support for NonNull in read queries (mutations are
not yet supported.)

The schemabuilder follows some simple rules to determine if a type is
NonNull or not:
- For scalar and struct types, a pointer type is nullable, and a
  non-pointer type is not.
- For list fields, all lists and their items are non-nullable. The
  executor already enforced that lists are non-nullable, and in
practice, our list items are never nullable.

Change the introspection test to use its own schema. Manually inspected
the schema and it looks reasonable.