batch: add a batching package
Batching is useful to reduce the number of RPCs issued by
queries such as

  {
    events {
      venue {
        address
      }
    }
  }

Naively, the venue resolver might issue an RPC for every single event.
Using batching, a batching venue resolver instead might issue a single
RPC for all events at once.

Batching is inspired by libraries like Facebook's dataloader, but
because Go has no promises, it instead uses time-based batching.
Experimentally, the extra latency can be as little as 200 microseconds
per call with reasonable results.

I attempted a different approach using explicit tracking of running
goroutines, but that felt brittle: Every goroutine would have to
carefully announce to the world when it is running and when not, and
that seems too easy to accidentally mess up, and hard to debug when
that happens. The API is mostly the same, and so the package could
later be changed to integrate with goroutine tracking, but for now,
this feels simpler.