graphql server: Limit number of concurrent rerunners
When a graphql server first starts up and has an influx of subscribers,
we've seen that it performs badly with a cold cache. This is because we
attempt to recompute the results of all 200 subscriptions concurrently,
which spreads the server's resources too thin.

In an effort to alleviate some of this pain, we are limiting the maximum
number of subscription rerunners that can execute in parallel to 64.
This is done by acquiring a semaphore in the NewRerunner computation
function and releasing the semaphore at the end of the function.