sqlgen: introduce intermediate BaseSelectQuery
Before this commit, there were two representations for queries: The
high-level implicit one, consisting of a sqlgen.Filter,
sqlgen.SelectOptions, and a result variable of type []*Foo. That got
immediately transformed into a sqlgen.SelectQuery, consisting of a
table name, set of columns, and fully baked sqlgen.SelectOptions.
Notably, the sqlgen.Filter got dropped.

This commit introduces a higher-level view that replaces the result
variable of type []*Foo with a reference to the table, but otherwise
keeps the passed-in Filter and SelectOptions. This is nice because it
lets other code operate on a higher-level view. For example, in the
livequery code, we were manually passing around a filter, but that is
now no longer necessary. Similarly, it will be helpful for query
batching to differentiate queries with empty SelectOptions from queries
with complex SelectOptions that might not be batchable.