graphql batch: Increase waiting period if batch fn is invoked rapidly
Adding support for a second timer in the batch function execution logic,
which is reset every time the batch fn is invoked. This is separate from
the MaxDuration timer, which is increased to 10ms as an absolute time
restriction. The interval timer will expire unless the batch fn is
invoked every 500us.