thunder errors: Replace custom errors with github.com/pkg/errors
Removing all of the custom code that dealt with wrapping errors and
generating stacktraces for errors, since the github.com/pkg/errors
package does the same in a better and more reusable way.

Note that calls to errors.New and errors.Errorf include a stacktrace of
the current stack in the reulting err, which is useful for any consumers
(like Raven) that want to evaluate the stacktrace at the time of the
panic.