reactive rerunner: Fix potential rerunner memory leak
Updating Rerunner.run() to immediately return if the cancelCtx function
is invoked, since we may have a client that has specified a large
minRerunInterval. For example, if the client specifies minRerunInterval
= math.MaxInt64, then run() will be called once with delay=0, and then a
second time (due to computation.node.handleInvalidate(r.run)) with a
very large delay. The second sleep causes the rerunner object and its
associated cache to stay alive for the lifetime of the process.

The fix for this potential memory leak is to select on both ctx.Done and
the sleep timer, and return immediately if cancelCtx() is called.