schemabuilder: remove buggy cache check in getType
For context, we cache the type of the known structs that we
have already seen so that we do not need to do reflection checks
on structs multiple times. However, this cache check prevents the
rest of the function from correctly wrapping non-nullable fields
in graphql.NonNull{}.

In practice, this bug is visible when a resolver returns a non-pointer
struct type of a struct type that we have already seen.

In this commit, I've removed the cache check. Note that the buildStruct
function still has this same cache check repeated the beginning of its
body, so we are still avoiding the penalty of re-reflecting on seen structs,
but we are allowing getType to run its null vs. non-null logic.