concurrencylimiter: bound number of concurrently running resolvers
The new concurrencylimiter package lets the graphql executor limit the number
of resolvers that are running at the same time. It integrates with the batch
package to allow resolvers that are joining a batched call to unblock other
resolvers.

All of this is opt-in using the concurrencylimiter.With call. Otherwise
everything behaves like before.