graphql: add directive parsing
Add parsing for directives on selections and fragments. Note that
the unused fragment code changes because we cannot rely on pointer
equality for checking global fragments - they may be used in multiple
locations in the document, so we store a copy with arguments per location
instead of sharing global versions.