sqlgen: support custom types
Our past support of custom types was suspect at best. I've added a test
for a not otherwise supported type ([16]byte).

Not attempted:
- Testing binlog (we shouldn't have trouble getting anything out,
  though, given that it uses BuildStruct under the hood)
- GraphQL (these changes are local to SQL only)

Caveats:
It seems like we need to implement `Value()` on the type without
indirect as we can pass non-pointers into SQL statements and the Valuer
interface would otherwise not be satisfied.