go sqlgen: support nullable (pointer) sql.Scanners
Currently, the only way to make sql.Scanner nullable is either by using
the sql.NullBool pattern (where we break up the value into `IsValid` and
`Bool`/whatever type).

This is because the single de-reference method can't overwrite the
pointer to be a null pointer. We also can't define two methods (one for
single de-reference and one for double de-reference), so the developer
story starts looking a little messy.

In order to support nullable custom types (for real), we will handle
`nil` on behalf of the scanner if it's a pointer.