go sqlgen: add non-ptr proto to benchmark
What we see here is an additional allocation for the non-ptr proto to
the heap. We expect this, since we have to create a new pointer in order
to satisfy the marshaler interface.

We deem this cost worth the additional safety of not having to deal with
pointers. Since people have to opt-in to use this feature there isn't a
noticeable difference to existing code-paths.