go schemabuilder: enforce key must be scalar
Key is used to match objects in diff & merge. We require keys
to be comparable in golang. If we serialize and then deserialize
a struct key without strong typing, it becomes map[string]interface{},
which is no longer comparable. Therefore, we want to enforce
a stronger requirement that keys are scalar, such as ints and strings.