schemabuilder: Nodes on connections are marked as non-null
If you have a list of pointers, we can assume that you still want to
return the values you're pointing to.

Alternatively, if we wanted to support nullable values in pagination
(which I don't think we do), we could make this an option. I think it's
ok to have an opinion here and to not allow this to happen unless we
are given a use-case down the road.