schemabuilder: Fix ptr-bug in paginated sorts/filters
When pointers are passed into buildFunction it gets things wrong.

When determining if the function is a ptr-func, it checks to see if the
ptr representation of a type matches what is passed in as an argument to
the function.

Because we are passing in a pointer, that pointer representation ends up
being a dual pointer.

Added an error so that we can't shoot ourselves in the foot this way
again.