go schemabuilder/build: Move top-level "build" funcs to build.go
Summary: Moves the main build functions (getType + enum+scalar) to their
own file.  This was mostly to isolate things out of the reflect package
but an argument can be made to put these into the "output" file (or move
that file into this one). I have no strong reasoning why to do it one way
or another.