sqlgen/db: change WithTx and WithExistingTx to return non-nil Context on error
These methods were returning a nil Context when they returned an error, but
that made these methods hard for callers to use the context when logging
the error. (They would have had to stash the original Context, but that
was too unobvious for callers to reliably do.)

This commit doesn't add tests because there is no db_test.go file.