go schemabuilder: Support reading SelectionSet in fieldfunc
Summary: Noticed this while writing tests for the batchfieldfunc.  We've
never passed in the SelectionSet properly for FieldFuncs, we'd always
return "nil" which probably caused some confusion for a bunch of
engineers.  This commit removes the selectionSet attached to the
funcContext and instead uses the passed in selectionSet from the Resolve
function. (It's possible this is not the functionality we want, please
comment if so).