go schemabuilder: Add (Dummy) BatchFieldFunc entrypoint for FieldFuncs.
Summary: Adds a BatchFieldFunc entry point for the Schemabuilder.
BatchFieldFuncs take a map[int]*ParentObject parameter as their "source"
and will return an equivalent response of map[int]*RespObject.  This
will allow us to reduce goroutine generation by batching together
fieldFunc queries.

This version is a "Dummy" version of the endpoint.  The Resolver we
return runs the BatchFunc with a single value, but all the code around
it is built to support passing in a slice of requests at once
(and returning a list of responses).