go graphql: Add support for executing batch FieldFuncs to new Executor
Summary: Adds entrypoints to the existing graphql.Field which will allow
us to add a "Batch" version of functions.  We include a way of flagging
whether an execution should be run as "Batch" or not.  In order to
guarantee that we don't change batch we set whether we're using batching
at the PrepareQuery stage.