go schemabuilder: Allow marking batchFieldFuncs as "NonNullable"
Summary: Adds support for marking a BatchFieldFunc as "NonNullable". We
don't have the ability to enforce this explicitly through the function
signature, but we can enforce it by validating that responses are being
passed back properly for every node we passed in.  If we see a mismatch
we fail the entire query.

This allows us to have parity with the current non-nullable fieldFuncs
we have right now.  I've added some tests to validate that we fail if a
nil result is returned, and for the happy cases of having non-nil
strings and enum values.