go reactive: Reset reactive cache on every run.
Summary: Whenever we get a retry sentinel error we want to throw away
the current calculations we have to get a clean slate for the next run.

This is required to rollout the new graphql Executor which has a
possibility of caching promises that stored an error (instead of the
result we want).