sqlgen: WithShardLimit bug when filter is part of pk
When we use UpdateRow, we fail to check if the shard filter is in the
where columns. If the shard is part of PK, the column and value is
added to as part of the columns in the where component of the query.
As a result, we should check both the where and the regular columns
for the shard filter when we update a row.