go batch: Make batch.Index a struct type
Summary: It's too easy right now to accidentally make a batch.Index from
a device Id or equivalent.  This commit changes the batch.Index to be a
struct type which cannot be explicitly created, you must use a function
to create it.  This should help reduce all the invalid usages of the
batch.Index.