go executor: Switch Batch check to call UseBatchFunc
Summary: There were some issues with passing the information about
whether to batch a function in the selection.UseBatch parameter.
Sometimes it just wouldn't get set for unknown reasons (after the
UseBatch func was replaces with an "Always on" function).  Instead of
going through and finding deep hidden meaning in go structs, this commit
just changes the logic for determining batch execution to call into the
usebatchFunc function in the execution path.