go graphql/schemabuilder: expensive field func option
Add a field func option to mark a field func as "expensive".
Previously, Thunder would assume that any field func accepting a "context"
parameter must be expensive and so should be executed concurrently.
The new behavior is to assume that a field func is inexpensive
unless the Expensive option is set explicitly.