go graphql: Expose the ctx for work units
Summary: In order to support advanced features like adding metric tags
per fieldfunc, we need to expose the context of the work unit.
Unfortunately trying to enumerate the functionality we want for the ctx
(e.g. Value/WithValue) would not work for all cases where we actually need
to pass in a real context to other functions (and they return a new ctx
with the type added).  So this commit just exposes the ctx value.