go reactive: Cancel context after each reactive Run
Summary: Some pieces of code will hold onto the context waiting for it
to be done (usually in a goroutine).  This means that all the objects
that are stored on the context are held until the original context is
marked as "done".  This would not happen until the websocket/http
request finishes.  So for continous requests we have a leak where each
intermediary context is holding onto the entire cache stack (and
anything else on the context) until we reset the websocket connection.

This commit adds a cancel to context for each reactive.Run.  This should
allow all code that depends on the context.Done() to be able to clear
between runs.