go graphql/schemabuilder: batch fallback always expensive
Default to expensive execution for non-batched fallback function.
This protects developers who might accidentally omit the
Expensive option when converting a FieldFunc to BatchFieldFuncWithFallback.

If someone is taking the time to convert a FieldFunc to batched,
then the FieldFunc must have been expensive, and we don't want
to introduce regressions by accidentally making it non-expensive.