sqlgen: add `UpdateRowWithResult` to `DB` interface
As a user of `sqlgen`, I'd like to be able to control code-flow in
certain cases where the number of rows updated is zero. `UpdateRow`
doesn't expose the `sql.Result` returned by `execWithTrace`, so this
information is available to be exported by refactoring the logic in
`UpdateRow` into an unexported function, and using said function in
`UpdateRow` and the new function, `UpdateRowWithResult` to reduce code
duplication.