federation: pass optional args
Sometimes we want to pass down metadata for the query from the first request to the underlying requests. We want developers to be able to inject their own federated gql server with additional args for any information they would also need to pass down. In the framework level, this looks like creating an optionlArg struct that is passed to the execute grpc request.

There is also some refactoring to make certain methods public to minimize the amount of code written while writing your own custom gql server client