federation: add optional response metadata
When the gateway makes a grpc request to one of the executors, it can get an optional response metadata field back. The gateway gets the union of all these responses and returns it with the results. This allows us to send information such as all the dependecies accumulated that are used to make the query live