federation: expose error if the response is nil
The response is nil if there is an error. When this happens, we cant cast nil to the object type specified for results, but we want to check the second argument to make sure the query doesnt error out. This allows us to send the specific error back to the user.