federation: simplify api to create a federated object
We initially had a function called FederatedFieldFunc to create the object, but we needed to ensure that you couldn't add that to anything. It should only be used for shadow objects. When we create a object, we can tell the schema builder it is a ShadowObject to register it in the schema, but we can also also create a custom shadow object that allows us to write a function to construct the shadow object given the keys. This means that if the key is just an id, we can make a database call to fetch the full object insteadd of passing the entire object through grpc