graphql/schemabuilder: remove the error for when an extra arg is provided
In a federated world we may delete a arg before the
gateway knows not to send the field. During this time, we don't
want the queries to fail if an extra arg is sent

This is safe to remove since if we send an extra argument, it just gets ignored.
It should not break any existing queries.