federation: fix infinite recursion bug
Before since we were adding add fragments of duplicate fields
regardless of uniqueness, and since we call flatten on all fragments,
we entered infinite recursion in some cases where we had multiple
matching fields with fragments in the selection set.

This fixes that bug so those requests will not go on forever and
OOM the container anymore.