federation: remove repeated selection and fragment parsing
When we were merging aliases, if we saw a repeated selection or fragment, we were adding it to the list again. Now we use a dictionary to check we don't add the same selection or fragment more than once, which helps reduce the number of times we need to flatten and merge aliases.

Once we parse a flatten a fragment, we remove it from the list of fragments since its already been parsed, which further helps prevent repeated parsing