go federation: bug in default service selector
For fields on the object (eg name on the device), its not a federated key but it does exist on multiple servers since its on the model. We dont want the service selector to ignore it, so I removed that check. Ideally devs need to write their own service selector when they use it, so this is only an issue with the default implementation. I also added a test to catch this