graphql/batch_executor: rm caching for expensive work units
Caching was only previously enabled for fields labeled Expensive.
The code generating the cache key uses reflect.Comparable to see if the
source object for the field can be hashed and used in the cache key.

However, reflect.Comparable cannot determine if the underlying type of
an interface is comparable.

This means that a struct with an interface field will panic
when used in the cache key if the underlying type of the interface
is non-comparable.

e.g.

```
type Foo interface {
    Bar()
}

type Uncomparable []int

func (u Uncomparable) Bar() {}

type Baz struct {
    Field Foo
}

x := Baz{
    Field: Uncomparable([]int{1,2,3}),
}
// The following returns true. Using `x` in the cache key will panic.
// because the slice field cannot be hashed.
reflect.TypeOf(x).Comparable()
```

Since caching was only enabled for fields labeled expensive, I
think it's reasonable to disable the caching for consistency.
Note that for non-websocket use cases, the cache would likely
not be used anyways.

If caching turns out to be important enough to reimplement,
consider a hashing the sources in a way such that non-comparable
fields are handled.