go federation: remove unnecessary fragments from flatten
Currently, we are maintaining a list of all fragments that are not
applicable to the current type in flattenFragments. For each
selectionSet, we append the list of unused fragments to the selection
set's fragments. However, we never again use selectionSet.Fragments
again, even in child recursive calls, and do not need it to flatten
the rest of the selectionSet. Thus, it is safe to remove this
logic.

As a result of this logic, there was a bug: if multiple selections use the same
fragment, then we mutate the fragment's selectionSet in child recursive
calls. This results in a fragment.SelectionSet.Fragments array that
grows exponentially in size, and runtime grows exponentially as well as
we process that fragment array. Removing this logic fixes this bug.