graphql: add support for nullable enum types in schemaBuilder
When generating schema.json, we handle Enums only if they are
non-nullable. If we supply a *EnumType (i.e. a pointer to an Enum), we
end up defaulting to the SCALAR type for that enum on a query.

This commit adds support for nullable enums to our schema generation,
ensuring that enum types are reflected correctly in the schema even if
they are marked as optional through use of a pointer.