graphql/schemabuilder/schema: enforce unique type names
Introspection assumes that each type in the schema is unique.
Note that package names are not considered when comparing type names.
In other words, foo_package.SomeType and bar_package.SomeType
are considered to have the same name.

This causes issues with the type resolution as only one type
of each name can exist in the schema. For example, one might
define 2 fieldfuncs, one returning foo_package.SomeType and
another returning bar_package.SomeType. The introspection
query will incorrectly declare that both fieldfuncs return
the same type: one of foo_package.SomeType or bar_package.SomeType.

In order to avoid type collisions, we will fail to build the
schema if we detect duplicate types in the GraphQL schema.

We iterate over every object's registered fieldfuncs,
recursing through each fieldfunc's argument and return types to
collect a set of types. Then, we compare these types to the
set of enum types.

Note that there is an existing type name uniqueness check in
getType. However, this only checks return types and does not
correctly compare slice or pointer types. For example,
registering both foo_package.SomeType and []*bar_package.SomeType
will not throw an error. The test for the existing check has been
removed in favor of the new tests introduced in this commit.