go federation: add handling for empty keys for non-root plans
It is possible for a federation subplan to receive an empty set of keys (this happens when the root plan returns an empty selection and passes it off to a subplan). When this happens, currently the executor assumes the lack of keys means the plan is a root plan, and will try to execute without a federation key, leading to an error if the service does not know about the field.

This commit will explicitly pass in whether or not an execution plan is a root plan or not, and will early return with an empty result if the plan is not a root plan, and the set of federation keys is empty.