go federation: supplement introspection schema
This commit adds queryType, mutationType, directives, and interface support to the introspectionschema output when merging multiple federated schemas.

This makes the output of mergeSchemas to exactly match the output of an introspection query, allowing us to use a merged federated schema as if it was the result of an introspection query.