go sqlgen: Add support for mysql USE INDEX and FORCE INDEX
Added support for mysql's USE INDEX and FORCE INDEX clauses.
This allows the developer to specify one or more indexes that
the mysql optimizer should choose. For more information, check out
the official documentation:
https://dev.mysql.com/doc/refman/8.0/en/index-hints.html.

If both UseIndex and ForceIndex are specified, only ForceIndex will
be applied. It's invalid mysql syntax to include both FORCE INDEX
and USE INDEX against the same table in a query.