go federation: create introspection client
For a federated schema created from multiple smaller schemas, we still want to be able to serve introspection queries as if it was a single schema. Introspection queries are to be served by an introspection client, which implements the federation.Executor interface, and stores the introspection query result in memory. The bytes are to be populated by the schema syncer.