federation: add obj passing support
Sometimes its easier to define an object as a federation key rather than
re-fetching the object from the db. It can also be more cost efficient
if its difficult to fetch it from the database.

However, to support this, the selection set needs to recursively list
all fields under all objects that are selected as keys. This commit adds
that support and also a comment explanation