go graphql: pass args to sort and filter methods
Allow users to provide filter and sort fields that implement the
connection arguments of the FieldFunc.

This allows for pre-filtering on sort fields in the event that a sort
field has to be filtered before providing sort values.  It also provides
more flexibility in the filter fields and what type of action they can
take if the arguments are passed directly to them.