go schemabuilder: apply object options even if object is already registered
Currently, when we define an object more than once within a single gql schema, we will ignore any options defined after the first registration (which is pseudo-randomly determined when the schema is built). We want to be able to apply options for every time we register an object, so we can have some regulations on object option code.

Currently, the only use for object options are for federation, where we use it to define a FetchObjectFromKeys method. Ignoring options on repeat registrations means a lot of code gets dropped. We want to ensure that FetchObjectFromKeys only gets defined once per gql server.